+++ lib/Service/RuleAnnotation.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace OCA\Files_AutoRename\Service;
+
+enum RuleAnnotation: string {
+    case ConflictCancel = 'ConflictCancel';
+    case ConflictKeepBoth = 'ConflictKeepBoth';
+    case ConflictKeepBothIfDifferent = 'ConflictKeepBothIfDifferent';
+}